backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-multipart==0.0.9
PyPDF2==3.0.1
openai==1.30.1
python-dotenv==1.0.1
alembic==1.13.1
pydantic-settings==2.2.1
httpx==0.27.0
